LayoutTests/fast/scrolling/mac/keyboard-scroll-after-wheel-event-cancel-expected.txt

PASS endScrollY > startScrollY is true
PASS successfullyParsed is true

TEST COMPLETE
